Move CTRCryptoInputStream back to stream package
