+++ src/main/java/com/intel/chimera/stream/CTRCryptoInputStream.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.chimera.stream.input;
+package com.intel.chimera.stream;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -31,6 +31,9 @@ import javax.crypto.ShortBufferException;
 
 import com.intel.chimera.cipher.Cipher;
 import com.intel.chimera.cipher.CipherTransformation;
+import com.intel.chimera.stream.input.ChannelInput;
+import com.intel.chimera.stream.input.Input;
+import com.intel.chimera.stream.input.StreamInput;
 import com.intel.chimera.utils.Utils;
 
 /**
