+++ src/main/java/com/intel/chimera/stream/CryptoInputStream.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.chimera.stream.input;
+package com.intel.chimera.stream;
 
 import java.io.IOException;
 import java.io.InputStream;
