+++ src/test/java/com/intel/chimera/stream/PositionedCryptoInputStreamTest.java
@@ -23,8 +23,6 @@ import com.intel.chimera.cipher.CipherTransformation;
 import com.intel.chimera.cipher.JceCipher;
 import com.intel.chimera.cipher.OpensslCipher;
 import com.intel.chimera.stream.input.Input;
-import com.intel.chimera.stream.input.PositionedCryptoInputStream;
-import com.intel.chimera.stream.output.CryptoOutputStream;
 import com.intel.chimera.utils.ReflectionUtils;
 import org.junit.Assert;
 import org.junit.Before;
